pruning seems to be working
