+++ pydfnworks/pydfnworks/general/general_functions.py
@@ -205,5 +205,11 @@ def from_pickle(self, filename):
         Best if used with DFNWORKS(pickle_file = <filename>)
     """
     import pickle
-    tmp = pickle.load(open(filename, "rb"))
-    self.__dict__ = tmp.__dict__.copy()
+    print(f"--> Loading DFN from {filename}")
+    if os.path.isfile(filename):
+        tmp = pickle.load(open(filename, "rb"))
+        self.__dict__ = tmp.__dict__.copy()
+    else:
+        error = f"Error. Cannot find pickle file {filename}.\nExiting program.\n"
+        sys.stderr.write(error)
+        sys.exit(1)
